Implement `vctrs_type2()` dispatch in C for simple types
